requirements.txt
asyncio==3.4.3
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
h11==0.14.0
idna==3.7
load-dotenv==0.1.0
numpy==1.26.4
outcome==1.3.0.post0
pandas==2.2.2
protobuf==5.27.0
pycparser==2.22
pyotp==2.9.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing-extensions==4.12.1
tzdata==2024.1
upstox-python-sdk==2.4.1
urllib3==2.2.1
uuid==1.30
websocket-client==1.8.0
websockets==12.0
wsproto==1.2.0
